Fix SetRecords not working with no existing records
This commit fixes a bug where SetRecords will error out if there are no
matching DNS records to be found instead of creating new ones.

This bug is caused by Client.getDNSRecords always erroring out when it
cannot return anything, while SetRecords expects an empty slice to be
returned with no error.

The fix is to make getDNSRecords not error out when there's nothing to
return.

The other use of getDNSRecords, which is in DeleteRecords, has been
updated to check whether the returned slice is empty. This is the only
other use of getDNSRecords.